追加処理仮実装

+++ nanairo-reader-app/src/main/java/eu/nanairo_reader/data/dao/BaseDao.java
@@ -7,7 +7,7 @@ public interface BaseDao<ENTITY, KEY> {
 
 	List<ENTITY> findList(ENTITY parameter);
 
-	int add(ENTITY entity);
+	long add(ENTITY entity);
 
 	int update(ENTITY entity);
